+++ packages/block-library/src/search/index.php
@@ -210,6 +210,8 @@ add_action( 'init', 'register_block_core_search' );
  * Ensure that the view script has the `wp-interactivity` dependency.
  *
  * @since 6.4.0
+ *
+ * @global WP_Scripts $wp_scripts
  */
 function block_core_search_ensure_interactivity_dependency() {
 	global $wp_scripts;